test(client): test leave mpe room and forced disconnection feature
